Remove the remainder of the components_deprecated folder
