--- rerun_py/rerun_sdk/rerun/components_deprecated/rect2d.py
@@ -1,29 +1,0 @@
-from __future__ import annotations
-
-from enum import Enum
-
-__all__ = [
-    "RectFormat",
-]
-
-
-class RectFormat(Enum):
-    """How to specify rectangles (axis-aligned bounding boxes)."""
-
-    XYWH = "XYWH"
-    """[x,y,w,h], with x,y = left,top."""
-
-    YXHW = "YXHW"
-    """[y,x,h,w], with x,y = left,top."""
-
-    XYXY = "XYXY"
-    """[x0, y0, x1, y1], with x0,y0 = left,top and x1,y1 = right,bottom."""
-
-    YXYX = "YXYX"
-    """[y0, x0, y1, x1], with x0,y0 = left,top and x1,y1 = right,bottom."""
-
-    XCYCWH = "XCYCWH"
-    """[x_center, y_center, width, height]."""
-
-    XCYCW2H2 = "XCYCW2H2"
-    """[x_center, y_center, width/2, height/2]."""
